#455217 [C#テキスト]1章演習問題　
新しいプロジェクトに第1章の演習問題のコードを書き写しました。
