--- 章末問題第1章/Practice1.2-3/MyClass.cs
@@ -1,17 +1,0 @@
-﻿namespace Practice1._2_1 {
-    /// <summary>
-    /// クラス
-    /// </summary>
-    class MyClass {
-        public int X { get; set; }
-        public int Y { get; set; }
-    }
-    /// <summary>
-    /// 構造体
-    /// </summary>
-    struct MyStruct {
-        public int X { get; set; }
-        public int Y { get; set; }
-    }
-
-}
