+++ 章末問題第1章/演習問題第1章 remake/演習問題第1章問2-3/MyClass.cs
@@ -1,0 +1,17 @@
+﻿namespace 演習問題第1章問2_3 {
+    /// <summary>
+    /// クラス
+    /// </summary>
+    class MyClass {
+        public int X { get; set; }
+        public int Y { get; set; }
+    }
+    /// <summary>
+    /// 構造体
+    /// </summary>
+    struct MyStruct {
+        public int X { get; set; }
+        public int Y { get; set; }
+    }
+
+}
