+++ 章末問題第1章/演習問題第1章 remake/演習問題第1章問１/Program1-1-1,1-1-2,1-1-3.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Henkou;
+namespace 演習問題第1章問１ {
+    class Program {
+        //Practice1.1-1 どら焼きオブジェクトを生成するコードを書いてください。この時の商品番号は98、商品価格は"210円"としてください。
+        //Practice1.1-2 どら焼きオブジェクトの消費税額を求め、コンソールに出力するコードを書いてください
+        //Practice1.1-3 Productクラスが属する名前空間を別の名前空間に変更し、Mainメソッドから呼び出すようにしてください。ただし、MainメソッドのあるProgramクラスの名前空間はそのままとしてください。
+        static void Main(string[] args) {
+            Product wDorayaki = new Product(98, "どら焼き", 210);
+            Console.WriteLine("消費税額は" + wDorayaki.GetTax() + "円です");
+        }
+    }
+}
